move remaining bits from local to main readme
The goal originally was to move everything to the newly created
terraform setup as it was all mingled. Things aren't as set up
as we thought via the terraform yet and the split just confuses
things for now especially with cleanup.

There is a chance we move from terraform back to EKSCTL but git
history will give us that if needed.